requirements.txt
selene == 2.0.0rc3
pytest == 7.4.0
allure-pytest == 2.13.2
pydantic == 1.10.7
jsonschema == 4.18.4
curlify == 2.2.1